Add files via upload
Added a tutorial message that happen before the game start telling you the instruction.
Added a way for the user to see what level you are on.
Added a way for the user to see what score you have and added a way to gain score(Temporary while working on code).
Added a way to speed up the program. (Need to add an error check to prevent going into the negative speed.)
Added a way to manually increase level, score, and speed for checking to see if this work.(Need to remove or edit for final project.)
Added a button to move the piece down on the board.